Just added the last message and last updated time to the group chat cards. I also refactored the code and arranged it in separate files
