+++ lib/app/app_setup.locator.dart
@@ -16,7 +16,7 @@ import '../core/services/auth_service.dart';
 import '../core/services/chat_service.dart';
 import '../core/services/group_service.dart';
 import '../core/services/user_service.dart';
-import '../ui/screen/home_view_model.dart';
+import '../ui/screen/home/home_view_model.dart';
 
 final locator = StackedLocator.instance;
