+++ lib/ui/screen/home/home_view_model.dart
@@ -1,9 +1,7 @@
 import 'dart:async';
-import 'dart:html' as html;
 import 'dart:typed_data';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:image_picker/image_picker.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 import 'package:web_groupchat/app/app_setup.locator.dart';
@@ -13,9 +11,9 @@ import 'package:web_groupchat/core/services/auth_service.dart';
 import 'package:web_groupchat/core/services/group_service.dart';
 import 'package:web_groupchat/core/services/image_picker_service.dart';
 import 'package:web_groupchat/core/services/user_service.dart';
-import '../../core/model/chat.dart';
-import '../../core/model/chat_model.dart';
-import '../../core/services/chat_service.dart';
+import '../../../core/model/chat.dart';
+import '../../../core/model/chat_model.dart';
+import '../../../core/services/chat_service.dart';
 
 class HomeViewModel extends BaseViewModel {
   final _user = locator<UserService>();
@@ -25,25 +23,44 @@ class HomeViewModel extends BaseViewModel {
   final _chat = locator<ChatService>();
   final _imgPicker = ImagePickerService();
 
-  Uint8List? _userDp;
-  Uint8List? _groupDp;
-
-  Uint8List? get userDp => _userDp;
-
-  Uint8List? get groupDp => _groupDp;
-
-  // String get groupDpUrl => _group.groupUrl;
-
-  // setGroupUrl(){
-  //   _selectedGroup.dpUrl = _
-  // }
-
-  User? get currentUser => _auth.currentUser;
-
   final String createGroupDth = "createGroup";
   final String addToGroupDth = "addToGroup";
   final String joinGroupDth = "joinGroup";
   final String leaveGroupDth = "leaveGroup";
+
+  Uint8List? _userDp;
+  Uint8List? _groupDp;
+
+  Uint8List? get userDp => _userDp;
+  Uint8List? get groupDp => _groupDp;
+
+  User? get currentUser => _auth.currentUser;
+
+  String get username => _user.username;
+
+  bool get isAdmin => _group.isAdmin;
+
+  GroupChatModel get selGroup => _group.selectedGroup!;
+
+  bool viewGroupDetails = false;
+  bool viewProfile = false;
+
+  GroupChatModel? _selectedGroup;
+
+  GroupChatModel? get selectedGroup => _selectedGroup;
+  ChatType _groupType = ChatType.values.first;
+  ChatType get groupType => _groupType;
+
+  Stream<List<GroupChatModel>>? get groupStream => _group.groupStream;
+
+  TextEditingController groupNameController = TextEditingController();
+  TextEditingController groupDescController = TextEditingController();
+  TextEditingController chatController = TextEditingController();
+  TextEditingController newUserController = TextEditingController();
+
+  List<GroupChatModel> get publicGroups => _publicGroups;
+  List<GroupChatModel> _publicGroups = [];
+
   final Map<String, StreamSubscription<List<ChatModel>>> chatStreams = {};
   final Map<String, List<ChatModel>> chatSnapshots = {};
   List<ChatModel> currChats = [];
@@ -52,12 +69,6 @@ class HomeViewModel extends BaseViewModel {
     _group.openGroupStream();
     getPublicGroups();
   }
-
-  bool get isAdmin => _group.isAdmin;
-
-  GroupChatModel get selGroup => _group.selectedGroup!;
-
-  bool viewGroupDetails = false;
 
   void closeGroupDetails() {
     viewGroupDetails = false;
@@ -69,18 +80,10 @@ class HomeViewModel extends BaseViewModel {
     notifyListeners();
   }
 
-  ChatType _groupType = ChatType.values.first;
-
-  ChatType get groupType => _groupType;
-
   onChangeType(ChatType type) {
     _groupType = type;
     notifyListeners();
   }
-
-  GroupChatModel? _selectedGroup;
-
-  GroupChatModel? get selectedGroup => _selectedGroup;
 
   void resumeSignal(GroupChatModel group, [bool justPausing = false]) {
     chatSnapshots[_selectedGroup!.id!] = currChats;
@@ -111,31 +114,26 @@ class HomeViewModel extends BaseViewModel {
         notifyListeners();
       });
     }
-    // _chat.openChatStream(selectedGroup!.id!);
     _selectedGroup = group;
     _group.selectGroup = group;
     notifyListeners();
   }
 
   Future<void> sendMessage() async {
-    await _chat.sendChat(
-        selectedGroup!.id!,
-        ChatModel(
-          sender: _user.email,
-          text: chatController.text,
-          time: DateTime.now(),
-        ));
+    ChatModel chat = ChatModel(
+      sender: _user.email,
+      text: chatController.text,
+      time: DateTime.now(),
+    );
+    await _chat.sendChat(selectedGroup!.id!, chat);
+    _group.updateLastMessage(chat, _selectedGroup!.id!);
     chatController.clear();
   }
-
-  Stream<List<GroupChatModel>>? get groupStream => _group.groupStream;
 
   void setChatType(ChatType type) {
     _groupType = type;
     notifyListeners();
   }
-
-  bool viewProfile = false;
 
   void onViewProfile() {
     viewProfile = true;
@@ -149,8 +147,6 @@ class HomeViewModel extends BaseViewModel {
     }
     notifyListeners();
   }
-
-  // Uint8List file;
 
   void pickUserDp() async {
     _userDp = await _imgPicker.pickWebImage();
@@ -167,17 +163,6 @@ class HomeViewModel extends BaseViewModel {
     }
   }
 
-  TextEditingController groupNameController = TextEditingController();
-  TextEditingController groupDescController = TextEditingController();
-
-  TextEditingController chatController = TextEditingController();
-
-  String get username => _user.username;
-
-  List<GroupChatModel> get publicGroups => _publicGroups;
-
-  List<GroupChatModel> _publicGroups = [];
-
   void getPublicGroups() async {
     _publicGroups = (await _group.getGroups())!;
     notifyListeners();
@@ -189,8 +174,6 @@ class HomeViewModel extends BaseViewModel {
       barrierDismissible: true,
     );
   }
-
-  TextEditingController newUserController = TextEditingController();
 
   void showNewUserDialog() {
     _dialog.showCustomDialog(
@@ -226,7 +209,6 @@ class HomeViewModel extends BaseViewModel {
     _selectedGroup!.members!.removeWhere((element) => element == _user.email);
     await runBusyFuture(
         _group.addUserToGroup(_selectedGroup!.members!, _selectedGroup!.id!));
-    // _selectedGroup = null;
     _publicGroups = [];
     notifyListeners();
     getPublicGroups();
